Untangle Unicode for Atoms
Erlang has support for UTF-8 encoded atoms, but they are special events.
This half-step gives us the ability to decode both types of atoms
(deprecated latin-1 and modern utf-8) and returns the same kind of Atom
object to Python using either unicode (py2) or str (py3).

Encoding still only sends latin-1 encoded atoms for the moment, so we
don't break backwards compatibility with any deployed version of this
library. Once the entire world is updated to at least this version of
the library we can enable encoding UTF-8 atoms.

To ensure we don't mess up, the encoder will raise if you try to encode
something that isn't latin-1.